Clean up __init__, split classes into own files, minor code cleanup
