+++ worlds/spyro/options.py
@@ -1,6 +1,6 @@
+from dataclasses import dataclass
 from typing_extensions import final
 
-from dataclasses import dataclass
 from Options import Choice, DeathLink, PerGameCommonOptions, FreeText, Toggle
