+++ worlds/spyro/rules.py
@@ -1,7 +1,7 @@
 from typing_extensions import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from . import SpyroWorld
+    from .world import SpyroWorld
 
 
 def set_rules(world: "SpyroWorld"):
